requirements.txt
Flask==2.3.2
PyYAML==6.0
requests==2.31.0
logging-loki==0.1.7
prometheus-client==0.16.0

opentelemetry-api==1.16.0
opentelemetry-sdk==1.16.0
opentelemetry-exporter-otlp==1.16.0
opentelemetry-instrumentation-flask
opentelemetry-instrumentation-requests
